Parchment + Porting

+++ common/src/main/java/net/createmod/catnip/math/VecHelper.java
@@ -283,7 +283,7 @@ public class VecHelper {
 		}
 
 		// ----- adjust for fov -----
-		float fov = (float) ((GameRendererAccessor) mc.gameRenderer).catnip$callGetFov(ari, partialTicks, true);
+		float fov = ((GameRendererAccessor) mc.gameRenderer).catnip$callGetFov(ari, partialTicks, true);
 
 		float half_height = (float) mc.getWindow()
 			.getGuiScaledHeight() / 2;
